pyproject.toml
[project]
name = "gophergpt"
version = "0.1.0"
description = ""
authors = [
    {name = "Maanas Taneja"}
]
readme = "README.md"
requires-python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
dependencies = [
    "langchain>=0.3.25,<0.4.0",
    "langgraph>=0.4.8,<0.5.0",
    "langchain-openai>=0.3.22,<0.4.0",
    "langchain-tavily>=0.2.3,<0.3.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "streamlit>=1.49.1,<2.0.0",
    "fastapi>=0.118.2,<0.119.0",
    "uvicorn>=0.37.0,<0.38.0",
    "chromadb>=0.6.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"